+++ Assets/Scripts/Game/ScoreManager.cs
@@ -7,9 +7,9 @@ namespace ShootAR
 	{
 		[SerializeField] private Text scoreLabel;
 
-		public int Score { get; private set; }
+		public ulong Score { get; private set; }
 
-		public static ScoreManager Create(Text scoreLabel = null, int score = 0) {
+		public static ScoreManager Create(Text scoreLabel = null, ulong score = 0) {
 			var o = new GameObject(nameof(ScoreManager)).AddComponent<ScoreManager>();
 			o.Score = score;
 			o.scoreLabel = scoreLabel;
@@ -24,7 +24,7 @@ namespace ShootAR
 		/// Adds points to the score and updates the GUI.
 		/// </summary>
 		/// <param name="points">The amount of points to add.</param>
-		public void AddScore(int points) {
+		public void AddScore(ulong points) {
 			Score += points;
 
 			if (scoreLabel != null) {
